+++ R/ModelDiagram.R
@@ -1,3 +1,37 @@
+as_epiworld_diagram <- function(...) {
+  structure(
+    ...,
+    class = "epiworld_diagram"
+  )
+}
+
+#' @export
+#' @rdname epiworld-model-diagram
+#' @param x An `epiworld_diagram` object.
+print.epiworld_diagram <- function(x, ...) {
+  cat(x, sep = "\n")
+  invisible(x)
+}
+
+#' @export
+#' @rdname epiworld-model-diagram
+#' @param ... Additional arguments passed to [DiagrammeR::mermaid()].
+plot.epiworld_diagram <- function(x, ...) {
+  # Replacing the first line
+
+  if (requireNamespace("DiagrammeR", quietly = TRUE)) {
+    x <- gsub("^flowchart", "graph", x)
+    DiagrammeR::mermaid(x)
+  } else {
+    stop(
+      "The 'DiagrammeR' package is required to plot 'epiworld_diagram' ",
+      "objects. Please install it using install.packages('DiagrammeR')."
+    )
+  }
+
+
+}
+
 #' Model Diagram
 #'
 #' Functions described here are helper functions for drawing
@@ -46,7 +80,7 @@ draw_mermaid_from_data <- function(
   stopifnot_string(output_file)
   stopifnot_bool(allow_self_transitions)
 
-  if (output_file != "") {
+  diagram <- if (output_file != "") {
     draw_from_data_cpp(
       states,
       transition_probs,
@@ -56,8 +90,8 @@ draw_mermaid_from_data <- function(
 
     message("Diagram written to ", output_file)
 
-    diagram <- readChar(output_file, file.info(output_file)$size)
-    return(diagram)
+    readChar(output_file, file.info(output_file)$size)
+
   } else {
     diagram <- capture.output(draw_from_data_cpp(
       states,
@@ -66,8 +100,10 @@ draw_mermaid_from_data <- function(
       allow_self_transitions
     ))
 
-    return(paste(diagram, collapse = "\n"))
+    paste(diagram, collapse = "\n")
   }
+
+  as_epiworld_diagram(diagram)
 }
 
 #' @rdname epiworld-model-diagram
@@ -127,7 +163,7 @@ draw_mermaid_from_file <- function(
   stopifnot_string(output_file)
   stopifnot_bool(allow_self_transitions)
 
-  if (output_file != "") {
+  diagram <- if (output_file != "") {
     draw_from_file_cpp(
       transitions_file,
       output_file,
@@ -136,8 +172,7 @@ draw_mermaid_from_file <- function(
 
     message("Diagram written to ", output_file)
 
-    diagram <- readChar(output_file, file.info(output_file)$size)
-    return(diagram)
+    readChar(output_file, file.info(output_file)$size)
   } else {
     diagram <- capture.output(draw_from_file_cpp(
       transitions_file,
@@ -145,8 +180,10 @@ draw_mermaid_from_file <- function(
       allow_self_transitions
     ))
 
-    return(paste(diagram, collapse = "\n"))
+    paste(diagram, collapse = "\n")
   }
+
+  as_epiworld_diagram(diagram)
 }
 
 #' @rdname epiworld-model-diagram
@@ -165,7 +202,7 @@ draw_mermaid_from_files <- function(
   stopifnot_string(output_file)
   stopifnot_bool(allow_self_transitions)
 
-  if (output_file != "") {
+  diagram <- if (output_file != "") {
     draw_from_files_cpp(
       transitions_files,
       output_file,
@@ -174,8 +211,7 @@ draw_mermaid_from_files <- function(
 
     message("Diagram written to ", output_file)
 
-    diagram <- readChar(output_file, file.info(output_file)$size)
-    return(diagram)
+    readChar(output_file, file.info(output_file)$size)
   } else {
     diagram <- capture.output(draw_from_files_cpp(
       transitions_files,
@@ -183,6 +219,8 @@ draw_mermaid_from_files <- function(
       allow_self_transitions
     ))
 
-    return(paste(diagram, collapse = "\n"))
+    paste(diagram, collapse = "\n")
   }
+
+  as_epiworld_diagram(diagram)
 }